Rewrite GUI for android
Compressed GUI's functionality into a non graphical based method. The Jpanel, JFrame, and Swing are not compatible with android apps. If the visual of the GUI was important, I would translate into an Activity file, but  the guidance is not centered around maps but with measurements, so the GUI should be refered to as Guidance as all visual componens of GUI has been stripped from it.